+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - AZERTY keyboard layout uses French AZERTY
  - QWERTZ keyboard layout uses German T3
  - CascadiaMono is no longer the default font
+ - No longer using dock widgets, Instruments + Waveforms are now located to
+   the right of the pattern editor.
 ### Fixed
  - Issue with Instrument Editor not properly updating when changing instruments
  - Hitting enter in the sequence editor line edit opens the Wave editor